# hw19-maven/src/main/resources/openapi.yml
openapi: 3.0.0
info:
  title: Endpoint для приема входящих сообщений
  description: |-
    API для приема входящих сообщений
  termsOfService: http://localhost/
  version: 1.0.0
servers:
  - url: http://localhost:8081/api/v1
...
paths:
...
  /user/:
    get:
      tags:
        - users
      summary: Список список пользователей
      description: Получение списка контекстов
      operationId: getUsers
      responses:
        '200':
          description: Успешное выполнение
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '400':
          description: ошибка запроса
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Errors'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Errors'
    post:
      tags:
        - users
      summary: Новый пользователь
      description: Создание нового пользователя
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: Успешное выполнение
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: ошибка запроса
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Errors'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Errors'
  /user/{id}:
    put:
      tags:
        - users
      summary: Новый пользователь
      description: Создание нового пользователя
      operationId: updateUser
      parameters:
        - name: id
          in: path
          description: id пользователя
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: Успешное выполнение
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: ошибка запроса
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Errors'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Errors'
...
components:
  schemas:
...
    Errors:
      type: object
      properties:
        errors:
          type: object
          additionalProperties:
            example: Описание ошибки
            type: string
...
    UserRequest:
      type: object
      title: Пользователь
      required:
        - login
        - password
        - passwordConfirmation
      properties:
        name:
          title: Имя
          type: string
        login:
          title: login
          type: string
        password:
          title: пароль
          type: string
        passwordConfirmation:
          title: подтверждение пароля
          type: string
        accesses:
          title: список доступов
          type: array
          items:
            type: string
            enum:
              - ADMIN
              - USER
        scopes:
          title: список контекстов
          type: array
          items:
            type: string
    UserResponse:
      type: object
      title: Пользователь
      properties:
        id:
          title: id пользователя
          type: string
        name:
          title: Имя
          type: string
        login:
          title: login
          type: string
        currentScope:
          title: Текущая область видимости
          type: string
        accesses:
          title: список доступов
          type: array
          items:
            type: string
            enum:
              - ADMIN
              - USER
        scopes:
          title: список контекстов
          type: array
          items:
            type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT